09-langsmith/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain-anthropic==0.3.0
langgraph==0.2.55
langsmith==0.1.147
structlog==24.4.0
tenacity==9.0.0
